requirements.txt
# Basis-Dependencies
numpy>=2.1.0
opencv-python>=4.8.0
pillow>=9.0.0
dlib>=19.24.0
tensorflow>=2.13.0
mtcnn>=0.1.1

# Supabase
supabase>=1.0.0

# API und Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
jinja2>=3.0.1
python-dotenv>=0.19.0
fastapi-utils>=0.2.1
aiohttp>=3.9.0

# Testing und Entwicklung
pytest==6.2.5
pytest-cov>=2.0.0
black>=21.0.0
isort>=5.0.0
mypy>=0.910
pylint>=2.9.0

# Logging und Monitoring
python-json-logger>=2.0.0

# Database
pydantic>=1.8.2
alembic>=1.7.1
SQLAlchemy==1.4.23

# ML/DL
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
facenet-pytorch>=2.5.2 